- corrected indentation, ordering of methods - reshape test passes when casted to float64 - equivalence test doesn't pass

+++ tests/models/efficientnet/test_image_processing_efficientnet.py
@@ -121,10 +121,10 @@ class EfficientNetImageProcessorTest(ImageProcessingTestMixin, unittest.TestCase
         for image_processing_class in self.image_processor_list:
             image_processor = image_processing_class(**self.image_processor_dict)
 
-            rescaled_image = image_processor.rescale(image, scale=1 / 127.5)
-            expected_image = (image * (1 / 127.5)).astype(np.float32) - 1
+            rescaled_image = image_processor.rescale(image, scale=1 / 127.5, dtype=np.float64)
+            expected_image = (image * (1 / 127.5)).astype(np.float64) - 1
             self.assertTrue(np.allclose(rescaled_image, expected_image))
 
-            rescaled_image = image_processor.rescale(image, scale=1 / 255, offset=False)
-            expected_image = (image / 255.0).astype(np.float32)
+            rescaled_image = image_processor.rescale(image, scale=1 / 255, offset=False, dtype=np.float64)
+            expected_image = (image / 255.0).astype(np.float64)
             self.assertTrue(np.allclose(rescaled_image, expected_image))
